+++ lib/colors/colors.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+const orangePrincipal = Color.fromARGB(255, 254, 109, 42);
+const greyFont = Color.fromARGB(255, 248, 248, 249);