+++ test/check_animation_test.dart
@@ -1,12 +1,20 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:check_animation/check_animation.dart';
 
 void main() {
-  test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
+  testWidgets('CustomUploadWidget displays title and button',
+      (WidgetTester tester) async {
+    await tester.pumpWidget(MaterialApp(
+      home: CustomUploadWidget(
+        animation: const AlwaysStoppedAnimation(1.0),
+        title: 'Upload Title',
+        buttonText: 'Upload',
+        onButtonPressed: () {},
+      ),
+    ));
+
+    expect(find.text('Upload Title'), findsOneWidget);
+    expect(find.text('Upload'), findsOneWidget);
   });
 }